Implement placeSellOrder and execution of selling stocks

+++ src/main/kotlin/de/hsrm/mi/stacs/pepjekt/services/IInvestmentAccountService.kt
@@ -15,7 +15,7 @@ interface IInvestmentAccountService {
 
     fun buyStock(investmentAccountId: Long, stockSymbol: String, purchaseAmount: BigDecimal): Mono<InvestmentAccount>
 
-    fun sellStock(investmentAccountId: Long, stockSymbol: String, volume: BigDecimal): Mono<InvestmentAccount>
+    fun sellStock(investmentAccountId: Long, stockSymbol: String, volume: Double): Mono<InvestmentAccount>
 
     fun getInvestmentAccountPortfolio(investmentAccountId: Long): Mono<InvestmentAccountDTO>
